requirements.txt
django==3.0.8
django-crispy-forms==1.9.1
django-tinymce==3.0.2
yandex-geocoder==2.0.0